requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.10
django-debug-toolbar==3.2.4
django-debug-toolbar-force==0.1.8
django-environ==0.8.1
django-heroku==0.3.1
django-nine==0.2.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.6
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.19
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
